+++ addons/project/static/src/components/subtask_kanban_list/subtask_kanban_list.js
@@ -5,7 +5,7 @@ import { Component, onWillStart } from "@odoo/owl";
 import { useService } from "@web/core/utils/hooks";
 import { registry } from "@web/core/registry";
 
-import { Record } from "@web/views/record";
+import { Record } from "@web/model/record";
 import { KanbanMany2ManyTagsAvatarUserField } from "@mail/views/web/fields/many2many_avatar_user_field/many2many_avatar_user_field";
 import { Field, getFieldFromRegistry } from "@web/views/fields/field";
 import { standardWidgetProps } from "@web/views/widgets/standard_widget_props";
