+++ addons/web/static/src/views/calendar/calendar_common/calendar_common_popover.js
@@ -5,7 +5,7 @@ import { Dialog } from "@web/core/dialog/dialog";
 import { evaluateBooleanExpr } from "@web/core/py_js/py";
 import { is24HourFormat } from "@web/core/l10n/dates";
 import { Field } from "@web/views/fields/field";
-import { Record } from "@web/views/record";
+import { Record } from "@web/model/record";
 import { getFormattedDateSpan } from "@web/views/calendar/utils";
 
 import { Component } from "@odoo/owl";
